feat: add search results hook
