+++ packages/feeds-client/@react-bindings/hooks/search-state-hooks/index.ts
@@ -1,0 +1,1 @@
+export * from './useSearchResult';
