+++ sample-apps/react-native/ExpoTikTokApp/components/Search/ActivitySourceResultList.tsx
@@ -1,33 +1,14 @@
-import {
-  type SearchSourceState,
-  useSearchResultsContext,
-  useStateStore,
-} from '@stream-io/feeds-react-native-sdk';
+import { useSearchResult } from '@stream-io/feeds-react-native-sdk';
 import { ActivitySectionList } from '@/components/ActivitySectionList';
-import { useStableCallback } from '@/hooks/useStableCallback';
-
-const searchSourceSelector = (nextState: SearchSourceState) => ({
-  items: nextState.items,
-  isLoading: nextState.isLoading,
-  hasNext: nextState.hasNext,
-  error: nextState.lastQueryError,
-});
 
 export const ActivitySourceResultList = () => {
-  const searchSource = useSearchResultsContext();
-
   const {
     items: activities,
     error,
     isLoading,
     hasNext,
-  } = useStateStore(searchSource?.state, searchSourceSelector) ?? {};
-
-  const loadMore = useStableCallback(async () => {
-    if (!isLoading && hasNext) {
-      searchSource?.search(searchSource?.searchQuery);
-    }
-  });
+    loadMore,
+  } = useSearchResult();
 
   return (
     <ActivitySectionList
